requirements.txt
langchain==0.0.284
huggingface_hub
transformers
sentence_transformers
gradio
numpy
Pillow
langchain-community